pyproject.toml
[project]
name = "airflow-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow~=3.0.0", # 3.0.x 시리즈만 허용
    "apache-airflow-providers-postgres>=6.2.3",
    "requests>=2.32.5",
    "astro-sdk-python[postgres]>=1.7.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]